fix(uniffi): make telemetry poll envelope carry full payloads
The first pass of the unified telemetry sink shipped with a poll
queue that, for four of the six record variants, enqueued *just
the timestamp as a string* where the UDL docstring promised the
full variant-specific JSON. The TypeScript facade then cheerfully
cast that envelope to `TelemetryRecord`, so `record.frame`,
`record.event`, `record.decision`, `record.snapshot` were all
silently `undefined` at runtime. Great for compile-time vibes,
useless at runtime.

While at it, the Protocol/Mls arms silently dropped the record on
any `to_json()` failure, two TS docstrings described capabilities
that didn't exist (`metricsCadenceMs=null` disabling emission, the
queue being populated before `installTelemetrySink`), the
field-level enum `_ => DefaultVariant` mappings hid core/FFI
drift, and the Android `encodeMetrics` was `.toInt()`-ing u32/u64
counters straight into signed-Int wrap territory.

The fix:

- Derive `Serialize` on the FFI DTOs with
  `rename_all = "camelCase"` (and an explicit
  `WiFiDirect → "wifiDirect"` rename to match the iOS/Android push
  encoders) and route every envelope through a single
  `envelope(category, fields)` builder. Push and pull now share
  one shape — the TS `TelemetryRecord` union is honest again.

- Add `emit_serialization_failure` so `to_json()` / `to_string()`
  failures land as a `telemetry.error.<name>` extension record
  with the error message, not as silent loss. Telemetry that
  swallows its own failures defeats the point.

- Replace `_ => ScoreUpdated` / `_ => InitialSelection` with
  `tracing::warn!` fallbacks so new-core/old-FFI skew is visible.

- Android: `putDouble(.toLong().toDouble())` for the UInt counters,
  since bytes-sent on a long-running relay absolutely reaches 2^31
  and turning that into a negative JS number is unhelpful.

- TS docstrings aligned with reality, and `metricsCadenceMs` typed
  as plain `number?` instead of lying about a `null`-means-disable
  path that the FFI can't express.

- Re-export `RetryQueueStats` from `offline-protocol` so tests in
  the uniffi crate can construct a `MetricsFrame` without pulling
  `offline-protocol-reliability` in as a direct dep.

- Six new adapter tests pin the envelope shape per variant, the
  drop-oldest queue semantics, and the extension-fallback path.
  The existing protocol-event test is updated for the new
  `eventJson` field name.

+++ bindings/react-native/android/src/main/java/com/offlineprotocol/OfflineProtocolModule.kt
@@ -746,16 +746,19 @@ class OfflineProtocolModule(reactContext: ReactApplicationContext) :
     }
 
     private fun encodeMetrics(m: TransportMetrics): WritableMap = Arguments.createMap().apply {
-        putInt("packetsSent", m.packetsSent.toInt())
-        putInt("packetsReceived", m.packetsReceived.toInt())
-        putInt("bytesSent", m.bytesSent.toInt())
-        putInt("bytesReceived", m.bytesReceived.toInt())
+        // Counter fields (UInt/UInt32/UInt64 on the Rust side) go through
+        // putDouble to avoid Kotlin's signed-Int wrap at 2^31 — long-running
+        // relays accumulate byte/packet counts well past that boundary.
+        putDouble("packetsSent", m.packetsSent.toLong().toDouble())
+        putDouble("packetsReceived", m.packetsReceived.toLong().toDouble())
+        putDouble("bytesSent", m.bytesSent.toLong().toDouble())
+        putDouble("bytesReceived", m.bytesReceived.toLong().toDouble())
         putDouble("errorRate", m.errorRate.toDouble())
-        putInt("avgLatencyMs", m.avgLatencyMs.toInt())
+        putDouble("avgLatencyMs", m.avgLatencyMs.toLong().toDouble())
         m.rssi?.let { putInt("rssi", it.toInt()) }
-        m.bandwidthBps?.let { putDouble("bandwidthBps", it.toDouble()) }
+        m.bandwidthBps?.let { putDouble("bandwidthBps", it.toLong().toDouble()) }
         m.congestion?.let { putDouble("congestion", it.toDouble()) }
-        m.queueDepth?.let { putInt("queueDepth", it.toInt()) }
+        m.queueDepth?.let { putDouble("queueDepth", it.toLong().toDouble()) }
         m.batteryLevel?.let { putInt("batteryLevel", it.toInt()) }
         m.isCharging?.let { putBoolean("isCharging", it) }
         m.relayConnectionCount?.let { putInt("relayConnectionCount", it.toInt()) }
